split data.lua
